[ALPS07779043] aee: fix type errors
Fix the type errors reported by clang-r475365 (clang 16.0.1)

Errors:
drivers/misc/mediatek/aee/aed/aed-debug.c:692:33: error: format specifies type unsigned long but the argument has type u64 (aka unsigned long long) [-Werror,-Wformat]
drivers/misc/mediatek/aee/aed/aed-debug.c:705:33: error: format specifies type unsigned long but the argument has type u64 (aka unsigned long long) [-Werror,-Wformat]
drivers/misc/mediatek/aee/mrdump/mboot_params.c:670:5: error: format specifies type unsigned long long but the argument has type unsigned int [-Werror,-Wformat]
drivers/misc/mediatek/aee/mrdump/mboot_params.c:670:12: error: format specifies type unsigned long long but the argument has type unsigned int [-Werror,-Wformat]
drivers/misc/mediatek/aee/mrdump/mrdump_arm64.c:124:33: error: format specifies type unsigned long but the argument has type u64 (aka unsigned long long) [-Werror,-Wformat]
drivers/misc/mediatek/aee/mrdump/mrdump_arm64.c:125:33: error: format specifies type unsigned long but the argument has type u64 (aka unsigned long long) [-Werror,-Wformat]
drivers/misc/mediatek/aee/mrdump/mrdump_panic.c:415:42: error: format specifies type unsigned long long but the argument has type void * [-Werror,-Wformat]
drivers/misc/mediatek/aee/hangdet/hangdet.c:645:52: error: format specifies type long but the argument has type unsigned long long [-Werror,-Wformat]
drivers/misc/mediatek/aee/hangdet/hangdet.c:646:3: error: format specifies type long but the argument has type unsigned long long [-Werror,-Wformat]
drivers/misc/mediatek/aee/hangdet/hangdet.c:646:31: error: format specifies type long but the argument has type unsigned long long [-Werror,-Wformat]
drivers/misc/mediatek/aee/hangdet/hangdet.c:646:55: error: format specifies type long but the argument has type unsigned long long [-Werror,-Wformat]

MTK-Commit-Id: 72658cfb92e202cbd5a80e905d81e721232adf5f

CR-Id: ALPS07779043
Feature: Kernel SI Operation

+++ drivers/misc/mediatek/aee/mrdump/mrdump_arm64.c
@@ -121,8 +121,8 @@ void mrdump_arch_show_regs(const struct pt_regs *regs)
 	print_pstate(regs);
 
 	if (!user_mode(regs)) {
-		pr_info("pc : [0x%lx] %pS\n", regs->pc, (void *)regs->pc);
-		pr_info("lr : [0x%lx] %pS\n", lr, (void *)lr);
+		pr_info("pc : [0x%lx] %pS\n", (unsigned long)regs->pc, (void *)regs->pc);
+		pr_info("lr : [0x%lx] %pS\n", (unsigned long)lr, (void *)lr);
 	} else {
 		pr_info("pc : %016llx\n", regs->pc);
 		pr_info("lr : %016llx\n", lr);
